riscv: kvm: Fix coverity issue in vcpu_pmu.c

+++ arch/riscv/kvm/vcpu_pmu.c
@@ -403,7 +403,7 @@ int kvm_riscv_vcpu_pmu_snapshot_set_shmem(struct kvm_vcpu *vcpu, unsigned long s
 {
 	struct kvm_pmu *kvpmu = vcpu_to_pmu(vcpu);
 	int snapshot_area_size = sizeof(struct riscv_pmu_snapshot_data);
-	int sbiret = 0;
+	unsigned long sbiret = 0;
 	gpa_t saddr;
 	unsigned long hva;
 	bool writable;
